+++ src/main/java/com/davivienda/factoraje/domain/dto/calculate/PercentageDTORequest.java
@@ -1,5 +1,7 @@
 package com.davivienda.factoraje.domain.dto.calculate;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class PercentageDTORequest {
 
-    private double interestRate;
-    private double base;
+    private BigDecimal interestRate;
+    private BigDecimal base;
     private int days;
 }
